fix: updated cli to core, spelling, refactore if not cli
