+++ lib/formats/dotnet-core-parser.js
