+++ lib/index.js
@@ -7,7 +7,7 @@ var path = require('path');
 var parseNuspec = require('./nuspec-parser');
 var jsonManifestParser = require('./json-manifest-parser');
 var debug = require('debug')('snyk');
-var projectJsonFormatParser = require('./formats/dotnet-cli-parser');
+var projectJsonFormatParser = require('./formats/dotnet-core-parser');
 
 function determineManifestType (filename) {
   switch (true) {
@@ -15,7 +15,7 @@ function determineManifestType (filename) {
       return 'project.json';
     }
     case /project.assets.json$/.test(filename): {
-      return 'dotnet-cli';
+      return 'dotnet-core';
     }
     case /packages.config$/.test(filename): {
       return 'packages.config';
@@ -76,7 +76,7 @@ module.exports = {
     var job = new Promise(function parseFileContents(resolve, reject) {
       var installedPackages = [];
       switch (manifestType) {
-        case 'dotnet-cli': {
+        case 'dotnet-core': {
           debug('Trying to parse dot-net-cli manifest');
           projectJsonFormatParser(fileContent, tree);
           resolve([]); // skip installed dependencies parsing
@@ -153,7 +153,7 @@ module.exports = {
         }
       }
     }).then(function scanInstalled(installedPackages) {
-      if (manifestType !== 'dotnet-cli') {
+      if (manifestType !== 'dotnet-core') {
         debug('Located ' + installedPackages.length + ' packages in manifest');
         function injectPath(dep) {
           dep.path = dep.localPath
@@ -207,28 +207,24 @@ module.exports = {
       }
     })
     .then(function fetchNugetInformationFromPackages() {
-      if (manifestType !== 'dotnet-cli') {
-        // initiate collecting information from .nuget files on installed packages
+      var nuspecParserChain = [];
+      if (manifestType !== 'dotnet-core') {
+        // begin collecting information from .nuget files on installed packages
         debug('Trying to analyze .nuspec files');
-        var nuspecParserChain = [];
         for (var name in flattendPackageList) {
           var dep = flattendPackageList[name];
           debug('...' + name);
           nuspecParserChain.push(parseNuspec(dep));
         }
-        return Promise.all(nuspecParserChain);
-      } else {
-        return Promise.resolve();
       }
+      return Promise.all(nuspecParserChain);
     }).then(function processNugetInformation(nuspecResolutionChain) {
-      if (manifestType !== 'dotnet-cli') {
+      if (manifestType !== 'dotnet-core') {
         nuspecResolutionChain.forEach(function (resolution) {
           if (!resolution) return; // jscs:ignore
           debug('.nuspec analyzed for ' + resolution.name);
           nuspecResolutions[resolution.name] = resolution;
         });
-      } else {
-        return Promise.resolve();
       }
     }).then(function buildDependencyTree() {
       // .nuget parsing is complete, returned as array of promise resolutions
